pyproject.toml
[project]
name = "aliexpress"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
